% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_ekio_fonts}
\alias{check_ekio_fonts}
\title{Check Font Availability}
\usage{
check_ekio_fonts()
}
\value{
Named list with font information by platform.
}
\description{
Reports which fonts are likely available on the current system.
}
\examples{
check_ekio_fonts()
}
