% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekio_colors.R
\name{ekio_colors}
\alias{ekio_colors}
\title{Enhanced EKIO Color System}
\usage{
ekio_colors(palette = "all")
}
\arguments{
\item{palette}{Character string specifying which palette to return}
}
\value{
Named vector of colors
}
\description{
Extended version of ekio_colors with mapping palettes, scientific palettes,
and custom EKIO combinations
}
\examples{
# EKIO palettes
ekio_colors("modern_premium")

# Mapping palettes
ekio_colors("RdBu")
ekio_colors("viridis")

# Scientific palettes
ekio_colors("hokusai1")

# Custom EKIO combinations
ekio_colors("midnight_steel")

}
