% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_ekio.R
\name{gt_ekio_simple}
\alias{gt_ekio_simple}
\title{Simple EKIO Table Theme}
\usage{
gt_ekio_simple(data)
}
\arguments{
\item{data}{A gt table object}
}
\value{
A styled gt table object
}
\description{
Quick styling function for common table use cases with minimal footer.
}
\examples{
\dontrun{
mtcars |> head(5) |> gt() |> gt_ekio_simple()
}
}
