% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_fira_code}
\alias{import_fira_code}
\title{Import Fira Code Font Family}
\usage{
import_fira_code(install = TRUE)
}
\arguments{
\item{install}{Logical, whether to install the font}
}
\value{
Invisibly returns font family name if successful
}
\description{
Downloads and imports Fira Code font family from Google Fonts
}
\examples{
\dontrun{
import_fira_code()
}
}
