% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_lato}
\alias{import_lato}
\title{Import Lato Font Family}
\usage{
import_lato(install = TRUE)
}
\arguments{
\item{install}{Logical, whether to install the font}
}
\value{
Invisibly returns font family name if successful
}
\description{
Downloads and imports Lato font family from Google Fonts
}
\examples{
\dontrun{
import_lato()
}
}
