% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_roboto_condensed}
\alias{import_roboto_condensed}
\title{Import Roboto Condensed Font Family}
\usage{
import_roboto_condensed(install = TRUE)
}
\arguments{
\item{install}{Logical, whether to install the font}
}
\value{
Invisibly returns font family name if successful
}
\description{
Downloads and imports Roboto Condensed font family from Google Fonts
}
\examples{
\dontrun{
import_roboto_condensed()
}
}
