% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_palettes.R
\name{list_external_palettes}
\alias{list_external_palettes}
\title{List Available External Palettes}
\usage{
list_external_palettes()
}
\value{
Named character vector with palette descriptions
}
\description{
Returns names and descriptions of available external (non-EKIO) palettes.
}
\examples{
list_external_palettes()
}
