% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_ekio_c}
\alias{scale_color_ekio_c}
\alias{scale_colour_ekio_c}
\alias{scale_fill_ekio_c}
\title{Continuous Color Scale}
\usage{
scale_color_ekio_c(palette = "blue", reverse = FALSE, ...)

scale_colour_ekio_c(palette = "blue", reverse = FALSE, ...)

scale_fill_ekio_c(palette = "blue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character. Palette name (default: "blue").
See \code{list_ekio_palettes("sequential")}, \code{list_ekio_palettes("diverging")}
and \code{list_ekio_palettes("academy")} for options.}

\item{reverse}{Logical. If TRUE, reverses the color order.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}
}
\value{
A ggplot2 scale object
}
\description{
Apply sequential, diverging or academy palettes to continuous/numeric data.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = hp)) +
  geom_point(size = 3) +
  scale_color_ekio_c()

ggplot(mtcars, aes(wt, mpg, color = hp)) +
  geom_point(size = 3) +
  scale_color_ekio_c("stone")
}
\seealso{
\code{\link[=scale_fill_ekio_c]{scale_fill_ekio_c()}}, \code{\link[=list_ekio_palettes]{list_ekio_palettes()}}
}
