% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekio_scales.R
\name{scale_fill_ekio_c}
\alias{scale_fill_ekio_c}
\title{EKIO Continuous Fill Scale}
\usage{
scale_fill_ekio_c(palette = "modern_premium", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of EKIO palette to use}

\item{reverse}{Logical, reverse the palette order}

\item{...}{Additional arguments passed to scale_color_gradientn}
}
\value{
ggplot2 scale
}
\description{
ggplot2 continuous fill scale using EKIO palettes
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_ekio_c("academic_authority")

}
