% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_ekio_palette}
\alias{show_ekio_palette}
\title{Display a Palette}
\usage{
show_ekio_palette(palette, n = NULL, labels = TRUE)
}
\arguments{
\item{palette}{Character or vector. Either a palette name or a vector of
hex colors.}

\item{n}{Integer. Number of colors (used for interpolation on sequential
palettes).}

\item{labels}{Logical. Show hex codes as labels (default: TRUE).}
}
\value{
A ggplot2 object (invisibly)
}
\description{
Deprecated. Use \code{\link[=ekio_pal]{ekio_pal()}} in its place.
}
\examples{
ekio_pal("contrast")
}
\keyword{internal}
