% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_ekio_fonts}
\alias{check_ekio_fonts}
\title{Check Font Availability}
\usage{
check_ekio_fonts(detailed = FALSE)
}
\arguments{
\item{detailed}{Logical, whether to return detailed information}
}
\value{
List or data frame of font availability
}
\description{
Checks which EKIO fonts are available on the system
}
\examples{
check_ekio_fonts()
check_ekio_fonts(detailed = TRUE)

}
