% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_ekio_font}
\alias{get_ekio_font}
\title{Get Best Available EKIO Font}
\usage{
get_ekio_font(type = "primary", loaded_fonts = NULL)
}
\arguments{
\item{type}{Font type: "primary", "secondary", "mono", "condensed"}

\item{loaded_fonts}{List of loaded fonts (from load_ekio_fonts)}
}
\value{
Font family name
}
\description{
Returns the best available font from EKIO hierarchy
}
\examples{
fonts <- load_ekio_fonts()
get_ekio_font("primary", fonts)

}
