% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_ekio_palettes}
\alias{list_ekio_palettes}
\title{List All Available EKIO Palettes}
\usage{
list_ekio_palettes(category = "all")
}
\arguments{
\item{category}{Filter by category: "all", "ekio", "scientific", "categorical", "mapping"}
}
\value{
Character vector of palette names
}
\description{
Returns a comprehensive list of all available EKIO color palettes
}
\examples{
# See all available palettes
list_ekio_palettes()

# Filter by category
list_ekio_palettes("categorical")
list_ekio_palettes("scientific")

}
