% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekio_scales.R
\name{scale_color_ekio_c}
\alias{scale_color_ekio_c}
\alias{scale_colour_ekio_c}
\title{EKIO Continuous Color Scale}
\usage{
scale_color_ekio_c(palette = "modern_premium", reverse = FALSE, ...)

scale_colour_ekio_c(palette = "modern_premium", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of EKIO palette to use}

\item{reverse}{Logical, reverse the palette order}

\item{...}{Additional arguments passed to scale_color_gradientn}
}
\value{
ggplot2 scale
}
\description{
ggplot2 continuous color scale using EKIO palettes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = hp)) +
  geom_point() +
  scale_color_ekio_c("modern_premium")

}
