% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_all_ekio_palettes}
\alias{show_all_ekio_palettes}
\title{Show All Palettes}
\usage{
show_all_ekio_palettes()
}
\value{
NULL (invisibly). Prints palette information to console.
}
\description{
Lists all available palettes organized by type.
}
\examples{
show_all_ekio_palettes()
}
