% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_ekio_palette}
\alias{show_ekio_palette}
\title{Display EKIO Palette Visually}
\usage{
show_ekio_palette(
  name = "modern_premium",
  n = NULL,
  indices = NULL,
  show_names = FALSE
)
}
\arguments{
\item{name}{Palette name}

\item{n}{Number of colors to show}

\item{indices}{Custom indices to display}

\item{show_names}{Logical, whether to show color names/codes}
}
\value{
ggplot2 object
}
\description{
Creates a visual representation of an EKIO color palette
}
\examples{
\dontrun{
# Basic palette display
show_ekio_palette("modern_premium")

# Show specific colors
show_ekio_palette("modern_premium", indices = c(9, 7, 5, 3))

# Show with color codes
show_ekio_palette("categorical", show_names = TRUE)
}

}
